update detect and crop face

+++ mobnet_ssd_detect_and_crop_face.py
@@ -1,13 +1,14 @@
 import os
 import cv2
 import numpy as np
+import imutils
 from imutils.video import VideoStream
 
-# prototxt_path = 'face-detector-model/ssd-model/deploy.prototxt'
-# caffemodel_path = 'face-detector-model/ssd-model/res10_300x300_ssd_iter_140000.caffemodel'
+prototxt_path = 'face-detector-model/ssd-model/deploy.prototxt'
+caffemodel_path = 'face-detector-model/ssd-model/res10_300x300_ssd_iter_140000.caffemodel'
 
-prototxt_path = 'face-detector-model/mobnet-ssd-model/ssd-face.prototxt'
-caffemodel_path = 'face-detector-model/mobnet-ssd-model/ssd-face.caffemodel'
+# prototxt_path = 'face-detector-model/mobnet-ssd-model/ssd-face.prototxt'
+# caffemodel_path = 'face-detector-model/mobnet-ssd-model/ssd-face.caffemodel'
 
 model = cv2.dnn.readNetFromCaffe(prototxt_path, caffemodel_path)
 
@@ -19,12 +20,12 @@ model = cv2.dnn.readNetFromCaffe(prototxt_path, caffemodel_path)
 
 vs = VideoStream(src=0).start()
 
-def crop_face(count):
-    # Create frame around face
-    image = cv2.imread("dataset/edi/image_capture.jpg")
-    (h, w) = image.shape[:2]
-    blob = cv2.dnn.blobFromImage(cv2.resize(image, (300, 300)), 1.0, (300, 300), (104.0, 177.0, 123.0))
-    # blob = cv2.dnn.blobFromImage(cv2.resize(image, (300, 300)), 1.0 / 127.5, (300, 300), (127.5, 127.5, 127.5), swapRB=True, crop=False)
+def detect_face(frame, key):
+    if key == 32:
+        frame = cv2.imread("dataset/edi/image_capture.jpg")
+    (h, w) = frame.shape[:2]
+    blob = cv2.dnn.blobFromImage(frame, 1.0, (300, 300), (104.0, 177.0, 123.0))
+    # blob = cv2.dnn.blobFromImage(cv2.resize(frame, (300, 300)), 1.0 / 127.5, (300, 300), (127.5, 127.5, 127.5), swapRB=True, crop=False)
 
     model.setInput(blob)
     detections = model.forward()
@@ -39,14 +40,14 @@ def crop_face(count):
 
         # If confidence > 0.5, show box around face
         if (confidence > 0.5):
-            cv2.rectangle(image, (startX, startY), (endX, endY), (255, 255, 255), 2)
+            cv2.rectangle(frame, (startX, startY), (endX, endY), (255, 255, 255), 2)
             # cv2.imwrite('dataset/edi/image_face_box.jpg', image)
-            frame = image[startY:endY, startX:endX]
-            cv2.imwrite('dataset/edi/face' + str(count) + '.jpg', frame)
-            print("Image cropped successfully")
+            frame = frame[startY:endY, startX:endX]
 
-    os.remove("dataset/edi/image_capture.jpg")
-    return count
+    if (key == 32):
+        cv2.imwrite('dataset/edi/face' + str(count) + '.jpg', frame)
+        print("Image cropped successfully")
+        os.remove("dataset/edi/image_capture.jpg")
 
 count = 0
 while True:
@@ -54,10 +55,8 @@ while True:
     # to have a maximum width of 400 pixels
     frame = vs.read()
     flip_frame = cv2.flip(frame, 1)
+    frame = imutils.resize(flip_frame, width=540)
 
-    # frame = imutils.resize(frame, width=400)
-    # show the output frame
-    cv2.imshow("Frame", flip_frame)
     key = cv2.waitKey(1) & 0xFF
 
     # if the `spacebar` key was pressed, write the *original* frame to disk
@@ -66,7 +65,12 @@ while True:
         count += 1
         cv2.imwrite('dataset/edi/image_capture.jpg', flip_frame)
         print("Image captured!")
-        crop_face(count)
+        # crop_face(count)
+        detect_face(frame, key)
+    else:
+        detect_face(frame, key=0)
+
+    cv2.imshow("detections", frame)
 
     # if the `q` key was pressed, break from the loop
     if key == ord("q") or key == ord("Q"):
